Figuring how to move things to a scraping class...
I don’t know if its the correct approach, but i will dissect the
scraping test and move it to a class which has several different
methods to scrape metadata. The basic approach and initialize function
will scrape filename/creationdate/filesize and store it to a cache
directory. other class methods can then add more metadata to the json
file, namely ffprober media specs and a hash (probably xxhash since it
seems to be the fastest around)… best would be if metadata scraping
methods could be extended in „plugins“ which define what file
extensions they can analyze and then add their metadata to the basic
scan…